set owner, pre, post upgrade for owner
