+++ magic_bridge/ic/src/magic_bridge/lib.rs
@@ -1,10 +1,13 @@
 use ic_kit::candid::{candid_method, CandidType, Deserialize, Nat};
 use ic_kit::macros::*;
-use ic_kit::{ic, Principal};
+use ic_kit::Principal;
 use std::cell::RefCell;
 use std::collections::HashMap;
+use types::*;
 
-use crate::factory::*;
+mod factory;
+mod types;
+mod upgrade;
 
 // Magic contract has controll over all the wrapped asset canisters
 // Check the mapping between ethere contract and ic
@@ -37,18 +40,13 @@ async fn handler(eth_addr: Principal, tokenType: TokenType, nonce: Nonce, payloa
     // Ok(Nat::from(1))
 }
 
+#[cfg(any(target_arch = "wasm32", test))]
+fn main() {}
 
-#[pre_upgrade]
-fn pre_upgrade() {
-    let stable_magic_state = STATE.with(|s| StableMagicState(s.0.take()));
-    ic::stable_store((stable_magic_state,)).expect("failed to save magic state");
-}
+#[cfg(not(any(target_arch = "wasm32", test)))]
+fn main() {
+    use ic_kit::candid;
 
-#[post_upgrade]
-fn post_upgrade() {
-    STATE.with(|s| s.0.borrow_mut().clear());
-
-    let (stable_magic_state,): (StableMagicState,) = ic::stable_restore().expect("failed to restore stable magic state");
-
-    STATE.with(|s| s.0.replace(stable_magic_state.0));
+    candid::export_service!();
+    std::print!("{}", __export_service());
 }
